an Affine Arithmetic lib libaffa implemented to replace Interval Arithmetic.

+++ Circuit_F2_1.h
@@ -2,14 +2,15 @@
 #define _CIRCUIT_F2_1_H_
 
 #include <vector>
+#include <aa.h>
 #include <boost/numeric/interval.hpp>
 
 typedef boost::numeric::interval<double> Interval;
 
 using namespace std;
 
-const int SIZE_PARM_F2_1 = 7;
-const int SIZE_RT_F2_1 = 3;
+const size_t SIZE_PARM_F2_1 = 7;
+const size_t SIZE_RT_F2_1 = 3;
 const double SHIFTD = 0;    //change the shift distance of imag axis here
 #define IC_F2_1
 #define SplitMethod_CFBM
